CringeBot/pyproject.toml
[tool.poetry]
name = "cringebot"
version = "1.0.0"
description = "A discord bot for moderation."
authors = ["AirSheep <<jxpkaxyz@gmail.com>>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.8.10"
"discord.py" = "^1.7.3"
colorama = "^0.4.4"
discord-components = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
